requirements.txt
streamlit
pandas
numpy
joblib
matplotlib
seaborn
scikit-learn
sqlalchemy
mysql-connector-python
scipy